// widget/servers/node/tsp-output/@typespec/http-server-javascript/models/all/typespec/rest/resource.ts
// Generated by Microsoft TypeSpec

import { Widget } from "../../demo-service.js";

/**
 * Resource create or update operation model.
 */
export interface WidgetUpdate {
  weight?: number;

  color?: "red" | "blue";
}

/**
 * Resource create operation model.
 */
export interface WidgetCreate {
  weight: number;

  color: "red" | "blue";
}

/**
 * Resource deleted successfully.
 */
export interface ResourceDeletedResponse {}

/**
 * Resource create operation completed successfully.
 */
export interface WidgetResourceCreatedResponse {
  /**
   * The status code.
   */
  statusCode: 201;

  body: Widget;
}

/**
 * Paged response of Widget items
 */
export interface WidgetCollectionWithNextLink {
  /**
   * The items on this page
   */
  value: Widget[];

  /**
   * The link to the next page of items
   */
  nextLink?: string;
}